pyproject.toml
[tool.poetry]
name = "mnemosinev3-0"
version = "0.1.0"
description = "MuseumCMS"
authors = ["and-pad <andrestor@gmail.com>"]
license = "GNU"
readme = "README.md"
packages = [{include = "mnemosinev3_0"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1.5"
djongo = "^1.3.6"
pymongo = "^4.6.2"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
